backend_fer/pyproject.toml
[project]
name = "backend-fer"
version = "0.1.0"
description = "backend for fer machine learning algo integration"
authors = [
    {name = "locinone",email = "locinone@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "huggingface-hub (>=0.29.2,<0.30.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torch (>=2.6.0,<3.0.0)",
    "jsonify (>=0.5,<0.6)",
    "requests (>=2.32.3,<3.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)"
]

[tool.poetry]
packages = [{include = "backend_fer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
